+++ cpp/open3d/core/hashmap/TemplateHashmap.h
@@ -54,10 +54,10 @@ namespace core {
 
 template <typename Hash, typename KeyEq>
 std::shared_ptr<DeviceHashmap<Hash, KeyEq>> CreateTemplateDeviceHashmap(
-        size_t init_buckets,
-        size_t init_capacity,
-        size_t dsize_key,
-        size_t dsize_value,
+        int64_t init_buckets,
+        int64_t init_capacity,
+        int64_t dsize_key,
+        int64_t dsize_value,
         const Device &device) {
     if (device.GetType() == Device::DeviceType::CPU) {
         return CreateTemplateCPUHashmap<Hash, KeyEq>(
